requirements/base.txt
Django==1.6.2
South==0.8.4
Unipath==1.0
django-braces==1.4.0
django-crispy-forms==1.4.0
django-debug-toolbar==1.0.1
django-flashpolicies==1.5
django-guardian==1.2.0
django-model-utils==2.0.2
gevent==1.0
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
psycopg2==2.5.2
six==1.5.2
sqlparse==0.1.11
wsgiref==0.1.2